in dev

+++ src/od.rs
@@ -1,9 +1,10 @@
 pub const OBJ_NAME_MAX_SIZE: u8 = 64;
 pub const OBJ_NUMB_MAX_SIZE: u8 = 128;
-pub const OBJ_CLASS_NUM: u8 =  80;
-pub const NMS_THRESH: f64 = 0.45;
-pub const BOX_THRESH: f64 = 0.25;
+pub const OBJ_CLASS_NUM: i32 =  80;
+pub const NMS_THRESH: f32 = 0.45;
+pub const BOX_THRESH: f32 = 0.25;
 
+#[derive(Debug, Default, Clone, Copy)]
 struct ImageRect {
     left: i32,
     top: i32,
@@ -11,12 +12,14 @@ struct ImageRect {
     bottom: i32,
 }
 
+#[derive(Debug, Default, Clone, Copy)]
 struct ObjectDetection {
     rect: ImageRect,
-    prop: f64,
+    prop: f32,
     cls_id: i32,
 }
 
+#[derive(Debug, Default, Clone)]
 pub struct ObjectDetectList {
     id: i32,
     count: i32,
